Fix for issue #949 (#951)
Should fix issue #949 

For reference, the issue is that when loading an existing craft file
that use configure setups that aren't allowed in the current game due to
tech restrictions, the setups are (correctly) not available in the GUI,
but they are still present/active on the part. Easily reproducible in
the default profile by loading a craft file with a mk1 pod having a
water recycler in a brand new career save.

Did zero testing for side effects, this should work alright but one
never knows with the configure module...