Update ReStockPlus_Science.cfg & Add Squad_Science.cfg (#963)
| ReStockPlus_Science | Squad_Science |
|--------|--------|
| <img width="1005" height="698" alt="Snipaste_2025-10-24_13-19-26"
src="https://github.com/user-attachments/assets/d1bbd622-1c97-44b2-8756-5abf5a604509"
/> | <img width="539" height="441" alt="Snipaste_2025-10-24_13-58-19"
src="https://github.com/user-attachments/assets/ce6599ce-d2b1-492e-956a-fafda553418a"
/> |

# ReStockPlus

https://github.com/Kerbalism/Kerbalism/blob/ef685c06b665eccfd5d8e79a108c0f8dca7381de/GameData/KerbalismConfig/Support/ReStockPlus_Science.cfg#L80

The `Experiment Return Unit` should be rebalanced. According to the
title and description of this part, it should be able to store more
data, rather than the original minimum of 0.5MB and a maximum of 2MB.
This is far from enough for the large amount of data that needs to be
returned, so I adjusted the current capacity.

```
@PART[restock-sciencebox-stack-1]:NEEDS[RestockPlus,FeatureScience]:AFTER[KerbalismDefault]
{
	@mass = 0.5
	@cost = 2500
	%MODULE[HardDrive]
	{
		%dataCapacity = 500
		%sampleCapacity = 4
	}
}
```
The adjustments are as follows:
1. Increase the mass to 0.5t
2. Minimum storage capacity to 500MB, maximum to 2GB
3. Cost increased to 2500

# Squad_Science
```
@PART[ScienceBox]:NEEDS[FeatureScience]:FOR[KerbalismDefault]
{
	@mass = 0.35
	@cost = 1500
	@MODULE[HardDrive]
	{
		@dataCapacity = 250
		@sampleCapacity = 8
	}
}
```

Add similar configuration for `Experiment Storage Unit`. 
However, I found that the original configuration seems to have set the
capacity to 0. I wonder why this is considered? If it is a data return
unit, it should also have a capacity?