Small Change to allow Process Controller Configure Overrides (#955)
Just a small change that allows me to override the Configure method in
the process controller. Should have no effect on anything in the
project. This is going to allow me to finish the full systemheat
integration into kerbalism using another mod.

Also added tiny change that saves the last multiplier as a protected
member, and added that multiplier to the GUI so uses can see it.